Updated mapping for price if value is int

+++ src/main/java/com/ofs/model/Inventory.java
@@ -49,7 +49,7 @@ public class Inventory implements OFSEntity {
         this.setType((String) map.get("type"));
         this.setCompanyId((String) map.get("companyId"));
         this.setQuantity((int) map.get("quantity"));
-        this.setPrice((double) map.get("price"));
+        this.setPrice((int) map.get("price"));
         this.setName((String) map.get("name"));
         this.setDescription((String) map.get("description"));
